+++ src/main/kotlin/crimera/patches/twitter/timeline/banner/HideBannerPatch.kt
@@ -2,18 +2,22 @@ package crimera.patches.twitter.timeline.banner
 
 
 import app.revanced.patcher.data.BytecodeContext
+import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
+import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.extensions.InstructionExtensions.getInstructions
-import app.revanced.patcher.extensions.InstructionExtensions.removeInstruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import com.android.tools.smali.dexlib2.Opcode
+import crimera.patches.twitter.misc.settings.SettingsPatch
+import crimera.patches.twitter.misc.settings.fingerprints.SettingsStatusLoadFingerprint
 import crimera.patches.twitter.timeline.banner.fingerprints.HideBannerFingerprint
 
 @Patch(
     name = "Hide Banner",
     description = "Hide new post banner",
+    dependencies = [SettingsPatch::class],
     compatiblePackages = [CompatiblePackage("com.twitter.android")],
     use = false
 )
@@ -30,7 +34,21 @@ object HideBannerPatch : BytecodePatch(
 
         val loc = instuctions.first{it.opcode == Opcode.IF_NEZ}.location.index
 
-        method.removeInstruction(loc)
+//        TODO: make the register dynamic
+        val HIDE_BANNER_DESCRIPTOR =
+            "invoke-static {}, ${SettingsPatch.PREF_DESCRIPTOR};->hideBanner()Z"
+
+        method.addInstructions(loc, """
+            $HIDE_BANNER_DESCRIPTOR
+            move-result v0
+        """.trimIndent())
+
+        SettingsStatusLoadFingerprint.result!!.mutableMethod.addInstruction(
+            0,
+            "${SettingsPatch.SSTS_DESCRIPTOR}->hideBanner()V"
+        )
+
+//        method.removeInstruction(loc)
 
     }
 }