+++ src/main/kotlin/crimera/patches/twitter/timeline/live/HideLiveThreadsPatch.kt
@@ -2,6 +2,7 @@ package crimera.patches.twitter.timeline.live
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
+import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.getInstructions
 import app.revanced.patcher.patch.BytecodePatch
@@ -10,10 +11,13 @@ import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
+import crimera.patches.twitter.misc.settings.SettingsPatch
+import crimera.patches.twitter.misc.settings.fingerprints.SettingsStatusLoadFingerprint
 import crimera.patches.twitter.timeline.live.fingerprints.HideLiveThreadsFingerprint
 
 @Patch(
     name = "Hide Live Threads",
+    dependencies = [SettingsPatch::class],
     compatiblePackages = [CompatiblePackage("com.twitter.android")],
     use = false
 )
@@ -30,9 +34,18 @@ object HideLiveThreadsPatch :  BytecodePatch(
 
         val loc = instructions.first{it.opcode == Opcode.IGET_OBJECT}.location.index
         val reg = method.getInstruction<OneRegisterInstruction>(loc).registerA
-        method.addInstruction(loc+1,"""
-            const v$reg, 0x0
+
+        val HIDE_LIVE_DESCRIPTOR =
+            "invoke-static {v$reg}, ${SettingsPatch.PREF_DESCRIPTOR};->liveThread(Ljava/util/ArrayList;)Ljava/util/ArrayList;"
+
+        method.addInstructions(loc+1,"""
+            $HIDE_LIVE_DESCRIPTOR
+            move-result-object v$reg
         """.trimIndent())
 
+        SettingsStatusLoadFingerprint.result!!.mutableMethod.addInstruction(
+            0,
+            "${SettingsPatch.SSTS_DESCRIPTOR}->hideLiveThreads()V"
+        )
     }
 }